requirements.txt
Pillow
requests
colorama
prettytable
pymysql
wechatpy
xmltodict
flask
sqlalchemy
cryptography